@@ -1,24 +1,15 @@
 /** @type {import('tailwindcss').Config} */
-module.exports = {
+export default {
   content: [
-    "./pages/**/*.{js,ts,jsx,tsx}",
-    "./components/**/*.{js,ts,jsx,tsx}",
+    "./index.html",
+    "./src/**/*.{js,ts,jsx,tsx}",
   ],
   theme: {
     extend: {
-      transitionProperty: {
-        height: "height",
+      fontFamily: {
+        pixel: ["VT323", "monospace"],
       },
-      colors:{
-        pink: '#d7387e',
-        lightblue: '#46eafb',
-        grey: '#1f1e1f'
-      }
-    },
-    fontFamily: {
-      roboto: ["Roboto Mono", "monospace"],
-      poppins: ['Poppins', 'sans-serif']
     },
   },
   plugins: [],
-};
+}
